Perfect numbers and gigaseconds.
Small surprise: 1e9 is float
Large surprise: 1 != 1.0

Dates have quite a comprehensive library, but building a datetime needs
a timezone that was not quite clear how the preferred method to build.

Cond is odd with `dup` and `drop` mixed to it. That makes a lot of sense
after you get used to it, but for simple cases like this it feels like
boilerplate code.